% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmda.plot.metrics.R
\name{hmda.plot.metrics}
\alias{hmda.plot.metrics}
\title{Plot model performance metrics across a grid of models}
\usage{
hmda.plot.metrics(
  df,
  metrics = c("auc", "aucpr", "r2", "mcc", "f2"),
  criteria = "rashomon",
  n_models = 100,
  tolerance = 0.05,
  plot = TRUE,
  title = NULL
)
}
\arguments{
\item{df}{A data frame of class \code{"hmda.grid.analysis"} containing a column \code{model_ids}
and numeric metric columns.}

\item{metrics}{Character vector of column names in \code{df} to be plotted.}

\item{criteria}{Character. One of \code{"n_models"} or \code{"rashomon"} (default).}

\item{n_models}{Integer. Number of top rows to plot when \code{criteria = "n_models"}.}

\item{tolerance}{Numeric in (0, 1). Alternative to \code{n_models}. Selects all models within
a given percentage distance of the best value for each metric (direction-aware). You must specify either
\code{n_models} or \code{tolerance}, not both. tolerance is direction-aware.
For example, when metric is AUC, if the tolerance is set to 1\%, it selects models that
have AUC equal or lower than 99\% of the model with the highest AUC.}

\item{plot}{Logical. If \code{TRUE}, prints the plot.}

\item{title}{Character. Add title to the plot.}
}
\description{
Creates a line plot comparing multiple (maximize-type) performance metrics across a set of models.
The input data frame is typically the output of \code{hmda.grid.analysis} and must contain
a \code{model_ids} column and one or more numeric metric columns (e.g., \code{aucpr}, \code{mcc}, \code{f2}).

The function can either plot the first \code{n_models} rows (\code{criteria = "n_models"})
or include all models that achieve at least \code{tolerance} times the best value
for at least one metric (\code{criteria = "rashomon"}).
}
\examples{
\dontrun{
  # Example: Create a hyperparameter grid for GBM models.
  predictors <- c("var1", "var2", "var3")
  response <- "target"

  # Define hyperparameter ranges
  hyper_params <- list(
    ntrees = seq(50, 150, by = 25),
    max_depth = c(5, 10, 15),
    learn_rate = c(0.01, 0.05, 0.1),
    sample_rate = c(0.8, 1.0),
    col_sample_rate = c(0.8, 1.0)
  )

  # Run the grid search
  grid <- hmda.grid(
    algorithm = "gbm",
    x = predictors,
    y = response,
    training_frame = h2o.getFrame("hmda.train.hex"),
    hyper_params = hyper_params,
    nfolds = 10,
    stopping_metric = "AUTO"
  )

  # Assess the performances of the models
  grid_performance <- hmda.grid.analysis(grid)

  # plot the metrics of models that are within 95\\% of the best models
  # for each of the specified metrics
  hmda.plot.metrics(grid_performance,
                    criteria = "rashomon",
                    tolerance = 0.95,
                    metrics = c("auc", "aucpr", "r2", "mcc", "f2"))

}

}
\author{
E. F. Haghish
}
